registrar domicilio docente

+++ database/migrations/2021_06_01_142046_create__b__domicilio_table.php
@@ -16,7 +16,7 @@ class CreateBDomicilioTable extends Migration
         Schema::create('_b__domicilio', function (Blueprint $table) {
             $table->increments('id');
             //FK Alumnos
-            $table->integer('alumno_id')->unsigned()->nullable();
+            $table->bigInteger('alumno_id')->unsigned()->nullable();
             $table->foreign('alumno_id')->references('nia')->on('alumnos');
             //FK Docentes
             $table->bigInteger('docente_id')->unsigned()->nullable();
